pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twinkle-eval"
version = "1.1.0"
description = "🌟 高效且準確的 AI 模型評測工具"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Teds Lin", email = "tedslin@gmail.com"},
    {name = "Liang Hsun Huang", email = "lianghsunh@gmail.com"},
    {name = "Min Yi Chen", email = "cyc00518@icloud.com"},
    {name = "Dave Sung", email = "code@k1dav.fun"},
]
maintainers = [
    {name = "Teds Lin", email = "tedslin@gmail.com"},
    {name = "Liang Hsun Huang", email = "lianghsunh@gmail.com"},
    {name = "Min Yi Chen", email = "cyc00518@icloud.com"},
    {name = "Dave Sung", email = "code@k1dav.fun"},
]
keywords = [
    "ai", "llm", "evaluation", "benchmark", "machine-learning", 
    "natural-language-processing", "artificial-intelligence", "testing"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.11"
dependencies = [
    "numpy~=2.3.0",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
    "openai>=1.93.0",
    "httpx>=0.28.1",
    "fastparquet>=2024.11.0",
    "datasets>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]

[project.urls]
Homepage = "https://github.com/ai-twinkle/Eval"
Documentation = "https://github.com/ai-twinkle/Eval#readme"
Repository = "https://github.com/ai-twinkle/Eval"
"Bug Tracker" = "https://github.com/ai-twinkle/Eval/issues"
Changelog = "https://github.com/ai-twinkle/Eval/releases"

[project.scripts]
twinkle-eval = "twinkle_eval.cli:main"

[tool.setuptools]
packages = ["twinkle_eval"]

[tool.setuptools.package-data]
twinkle_eval = ["*.yaml", "*.yml", "config.template.yaml"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["twinkle_eval"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]